requirements.txt
FastAPI
discord.py
uvicorn
python-dotenv
requests
